Fix issue with tray icon double clicking.
When clicking the tray icon in rapid succession on
windows the next click on it wouldn't register.
Analysis showed the signal really didn't reach the
handler. It seems like there are some bad things
happening when calling activateWindow() in this
handler as when it is called this issue disappears.
As our setWindowState call should perform window
activation anyways we now simply rely on it and
remove the other call. Also unified both track
click handlers.

Fixes #1452