Remove unused injection code and function lookups
* Remove GL function resolution of not used functions
* Remove unused SwapBuffer function hooks:

TL;DR: SwapBuffers calls wglSwapBuffers, so we do not need to hook it,
and wglSwapLayerBuffers is for layer planes, which are optional and of
arbitrary number and index - so we never use that to hook/draw.

In detail:

SwapBuffers is actually a function of Gdi32.dll, and the one
youre supposed to use on windows (MSDN [1]), at least when
statically linking.
The wgl functions are used when dynamically linking.

See OpenGL Windows FAQ [2]

* 5.190 What do I need to know about mixing WGL and GDI calls?
* 5.210 How do I properly use WGL functions?

> So since GDI and the ICD share function names like SwapBuffers, a
> wglSwapBuffers is needed to avoid ambiguity when loading the proc
> dynamically.

Quoted from [3]

wglSwapLayerBuffers is the swap function for layer planes Overlay, Underlay,
and Main Planes. [4]

So using wglSwapBuffers is always enough.
We could draw on other layers, but if hooking into wglSwapLayerBuffers we
would have to know which layer to draw on, and either call it from
wglSwapBuffers or not, depending on the original program rendering that
layer, so thats not feasible.

In ET:L, for every logged GDI SwapBuffers call, a wglSwapBuffers call follows.

[1] http://msdn.microsoft.com/en-us/library/dd369060%28v=vs.85%29
[2] https://www.opengl.org/archives/resources/faq/technical/mswindows.htm
[3] https://www.opengl.org/discussion_boards/showthread.php/141384-SwapBuffers-or-wglSwapBuffers
[4] http://msdn.microsoft.com/en-us/library/dd368815%28v=vs.85%29