Check if the host supports 64-bit processes before attempting to start the x64 helper process.
This change adds a function, canRun64BitPrograms() to OverlayWinPrivate.

If the program is built for x64, the function unconditionally
returns true.

If the program is built for x86, the function uses a call to the
runtime-resolved kernel32!IsWow64Process to determine whether or
not the system is x64 capable.

This functionality is needed to avoid spawning the x64 helper in
case the system cannot support it. Doing so without care, like we
did previously, could be fatal because Mumble immediately restarts
the helper process if it terminates. This would presumably make the
host computer unusable.