GlobalShortcut_unix: disable evdev by default.
...and gate it behind the hidden 'shortuct/linux/evdev/enable'
config option.

We don't actively want people to use evdev unless
they really have to (as it requires people to change the
permissions of their keyboard device node to be quite wide-open
to be useful).

Our evdev code has also caused problems for users of
Yubikeys and other gadgets that identify themselves as
keyboards. Those problems should be a thing of the past
with this commit.

Fixes #1145