Fix typo in speex-build.pro: '../speex-build' instead of '../speex-srcbuild'.
The new 'no_include_pwd' CONFIG option introduced by
d855b67d10b493afbae7cb5756ecf0d1b6270722 showed us this bug.

Before that commit, OS X *did* include PWD in the INCLUDEPATH
by default. Because of that, the equivalent of '../speex-build'
was *already* in the INCLUDEPATH, because 'speex-build' is the
PWD of the .pro file.

Now that we explicitly require the PWD not to be included, our
own inclusion of 'speex-build' must be spelled correctly to work.

This commit fixes that.