Use non-negative error constants in overlay_exe.h.
The Win32 ExitProcess() API takes an UINT, and we
used to pass negative values to it.

Obviously, those negative values can be represented
just fine in an unsigned integer, but for the sake
of following the API and to avoid future confusion
this commit changes the error constants used by
the overlay helper program to be non-negative.