Move label to bottom of connect dialog and default it to hostname
This patch improves the usability of the connect dialog by
moving the server label to the bottom of the dialog and making
it track the hostname. This way the user gets a sensible default
while also being able to manually edit it afterwards.

Once the user edits the label it is seen as custom and will no
longer track the server address until the label field is
manually cleared by the user.