PulseAudio: Add an option to avoid attenuating applications on other outputs (sinks) that Mumble is not attached to (i.e. HDMI output or an external headset). Also, avoid attenuating module-loopback devices attached to Mumble's sink. Loopback modules are used to link sinks in advanced setups, and if an application was specifically moved to another sink to avoid attenuation but the output is fed back to Mumble's sink via a loopback, the loopback would be attenuated (it would be a sink input) leading to an indirect application attenuation.
Before this commit, Mumble indiscriminately attenuates applications. A simple example: it would attenuate applications that were set to output to HDMI -- something that wasn't always desirable. My HDMI wouldn't need to be attenuated because I am able to physically distinguish between sound coming from either my HDMI or my laptop and "tune in" accordingly.

A more advanced use case is my PulseAudio streaming setup (a diagram of which is available at http://208.68.38.147/PA_Config.odg). I create two additional sinks: "stream" and "stream_spkr." The "stream" sink is used when I want to send audio from an application to my stream but not to my speakers/headset (i.e., background music or something). In this case, before this commit Mumble would attenuate my program playing the background music. If I am running Mumble off-stream (which I often am), this leads the background music volume fluctuating for seemingly no reason.

The second sink, "stream_spkr" routes both to "stream" and my speakers/headset via two module-loopbacks. In this way, anything I attach to stream_spkr can be heard both by my viewers and myself.

As far as dodging module-loopback inputs, I can think of reasons to do this as well as reasons not to. It should be done in the case where a user explicitly moves an application to another sink that is module-loopbacked to Mumble's sink for the sole reason of avoiding attenuating for that program. However, in scenarios like my streaming setup, this wouldn't be ideal. Another option should probably be added to turn this on and off. However, adding another option would essentially fill out the options dialog, so I hesitate to do so.

I only supported PulseAudio in this patch, but this functionality could probably be extended to other audio systems