Add support for running with a versioned root path in mumble.exe.
This commit bakes the MUMBLE_VERSION string into mumble.exe.

This version string is used to determine where to load mumble_app.dll
from, and what to consider the root directory for other binaries.

In pseudo code, if the path

    ${MUMBLE_EXE_DIRECTORY}\Versions\${MUMBLE_VERSION}

exists, the mumble.exe binary will use that as the root for loading
mumble_app.dll, and all further binaries.

If mumble.exe finds the aforementioned directory, it sets the
MUMBLE_VERSION_ROOT environment variable to signal which directory
to use as the "application root" directory for loading further
binaries.