Remove old "non-tagged" Boost include directories.
The default Boost installation paradigm for non-Unix boxes
is to install Boost into your include directory with a version
attached. For example:

    include/boost-1_49/boost

However, it is also possible to install Boost in a more traditional
manner, such that it simply lives in include/boost, instead of having
a version attached -- this is called a tagged installation, and
is triggered via 'b2 --layout=tagged'.

All of our build environments use tagged Boost installations, and thus
we can avoid the embedded versions now that we have transitioned to
using build environments from mumble-releng.

While we don't necessarily want to force the build environments
from mumble-releng on people, the versioned includes are a bad idea
in the first place, because they require constant maintenance.

Nothing prevents people from adding a non-tagged include themselves,
if that is what their home-grown build environment uses.