Suppress protobuf warnings in unix builds
The protobuf header creates lots of spurious warnings
throughout our compile which we can do nothing about.
To get rid of this noise this patch makes protobuf
generate into a separate directory which we mark as a
system include directory (-isystem) on unix. This way
we should not longe receive warnings out of it.