Enable audio to be played during config restart query
The ConfigDialog::apply function also re-initializes
the audio input by stopping it at is beginning and
restarting it at the end. The modal restart query dialog
blocked this audio restart leaving the user with no
audio until the dialog was dismissed. This patch moves
the query until after the audio restart to resolve this.