Add ThemeInfo class for finding and loading Theme descriptions
Adds the ThemeInfo and ThemeInfo::StyleInfo classes which
describe a Mumble theme with multiple styles. The basic functionality
consists of scanning a directory for theme directories containing
a theme.ini description and parsing that. A fall-back mechanism checks
directories missing a theme.ini for qss files and constructs a virtual
theme with corresponding styles for it. This will enable users to keep
using most existing styles unchanged by simply extracting them into a
themes directory.

For more information see the documentation contained in the classes
themselves.