Add built-in "Metro Mumble" theme and set "Lite" variant as default.
Adds the Metro Mumble theme as a submodule and ensures it is built
into the binary itself as a resource. Also changes the default
configuration to select the Mumble Metro theme in it's Lite style
as the default. Whether we go with lite or dark in the end we'll
want to select a Metro Mumble style as the default. This will allow
us to always switch to the other without restart because they use
the same icons.