Add generate-qrc.py script as a rcc -project replacement
One can use rcc -project to emit .qrc files for the contents of
a directory. However this functionality seems to be woefully incomplete
to the point of being broken.

This small python script offers the same functionality with more
options to customize and configure behaviour. This includes custom
prefixes, better handling of qrc file path vs. alias and exclusions.