Fixup: Suppress protobuf warnings in Unix builds
Note: Previous way wasn't viable as isystem doesn't affect
      includes directly next to the file which happen as the
      .cc file includes the header next to it.

Run protoc and generated code compilation as a seperate
mumble_proto target producing a static lib. This lib is then
linked by mumble and murmur.

The folder of this module is added to the include paths so
Mumble.pb.h is available. On Unix platforms a -isystem entry
is added for the directory.

The advantages of this approach is that we can disable warnings
for the compilation of this target like we do for 3rd party modules.
Additionally we no longer generate and compile the protobuf files
twice.