Work around Windows permissions preventing Mumble restart
Mumble fails to restart itself when asked to do so from the configuration menu.
This failure is handled correctly and the shutdown in general is clean.

This is because we run with uiaccess=true resulting in the normal
QProcess::startDetached failing with the following in the windows
error log: "The process failed to handle ERROR_ELEVATION_REQUIRED
during the creation of a child process."

QProcess doesn't correctly consider windows permissions when performing
launch operations as can be seen in QTBUG-7645 so we have to work
around this ourselves. As we do not need full admin privileges and
do not want the user to be prompted use the ShellExecuteEx function
which transparently acquires the needed permissions.

Fixes #1747