Ensure closeEvent for MainWindow is called on self-restart
The previous code triggered the restart of Mumble directly from
the ConfigDialog apply function by calling qApp->exit with the
magic exit code. This does not call the MainWindow closeEvent which
handles storing MainWindow geometry.

This patch moves the actual restart query and execution from the
ConfigDialog to the MainWindow. It is executed once the apply of
the ConfigDialog was run and triggers a restart if needed. To do
so the window has a new flag 'restartOnQuit' which the closeEvent
can pick up.

Fixes #1716