Fix setting an empty SuperUser password not creating an error
Mistakenly checked string for empty instead of null resulting
in a set but empty string not causing the correct error message.
Instead the request to set a password was ignored and the server
started normally. This patch resolves this issue. If the password
is set but empty an error is emitted and the server isn't started.

Fixes #1778