Enable LookConfig to detect changes to user themes folder
LookConfig now uses a QFileSystemWatcher to react to changes
to the user themes folder. This way changes are directly visible
in the themes combobox. To prevent flooding with change events
when copying in a full directory we defer the update until 200ms
have passed after the last change we were notified of.