Replace all comments using hash (#) w/ semicolons (;)
Lines starting with a hash are not considered to be comments!!

The QSettings() class has no formal support for comments. In fact,
there's no mention of comments at all in the class documentation:
http://doc.qt.io/qt-5/qsettings.html

There is some limited support for comments by denoting a line with a
semicolon. You can confirm this via the associated source code:
https://github.com/qtproject/qtbase/blob/5.6/src/corelib/io/qsettings.cpp

However, if saving the file via the Qt interfaces, comments will
generally be stripped out. This isn't to my knowledge a problem for
Murmur as there's no case where the server itself will update its
configuration and save the changes back to its INI file automatically.

The existing sample INI file prior to this commit only ever worked as
there's an even number of unescaped special characters in the header!