Use lookup table for cipher info in 'Connection Info' dialog.
This patch introduces a lookup table that allows us to retrieve
additional IETF TLS parameters based on the ciphersuite name
provided by Qt/OpenSSL. With this additional information we can
have the detailed output we want to have for the connection
info dialog.

The table is generated by the generate-cipherinfo.py script
which uses a heuristic to select a representative set of
suites we might to expect to see from the the official IETF
TLS parameter descriptions.

Should we not find the cipher suite a connection ended up
using in this table we will fall back to less detailed output
which the user can use to find the specific parameters.

This patch also contains some other minor changes to the dialog.