Make overlay FPS and clock positionable
* Implement #1068, making the FPS and clock in the overlay positionable
like our user-block

* Introduce class OverlayPositionableItem to generalize our red
draggable anchor to set the position in the overlay settings
** Has a configuration and a display mode; with and without the anchor
** Receives a pointer to the position setting

The fps position was already set up to save, the timer not.
Use storage names according to the show flags ("time" rather than
"clock").

The anchor ("handle") and text-item are added to the scene as independant
objects to allow moving the anchor to the bottom/right, at which point
the item is prevented from bleeding over the edge (so a linked movement
of individual items with custom behavior at the right/bottom edge case).

Additional Notes:

Using relative values for positioning, the default values are best-effort
values. The space between the items (the top value of the FPS) will differ
between different screen/scene sizes.

The position is saved as a rectangle, even though a point would be enough.

Handling of visibility is sub-optimal, but works. (As we use two
independent, linked items, we have to implement our own setVisible
method, and currently only call in OverlayConfig (when using config
mode).

It looks like the overlaytext used to create a text pixmap could be
replaced by QGraphicsTextItem.