Auto-generate qrc file for embedded Qt translations when running qmake.
This change removes our qmake-based Qt translation embedding.

That system uses mumble_qt.qrc resource file with
hardcoded filenames for Qt translations, and some logic
implemented in qmake that copies Qt translations into
the Mumble source tree such that the paths in the
mumble_qt.qrc file match.

The new system introduces a simple Python script that
takes an output filename for the .qrc file the tool
will write, along with a set of directories containing
Qt translations.

The tool will generate a Qt resource file containing
references to all the translation files found in the
specified directories. However, the tool takes care
to only include language files once.

In typical use, the first directory parameter passed
to the tool is the QT_INSTALL_TRANSLATIONS directory,
which is where Qt stores its own translation files.
The second directory is Mumble's fallback directory.

The tool then goes through all files in the first
directory, and notes down which languages have been
processed. Multiple files for a single langauge can
be included from the a directory (qt_help_da.qm,
and qtbase_da.qm), but once a language has been
added from one directory, it will not be added
if found in the next one in line.

We use this to include a set of 'fallback'
translations for versions of Qt that do not
include them. This also allows this new style
of Qt translation embedding to be forward
compatible with newer versions of Qt that
add new translations.

Once Qt includes a translation that we have
in our fallback directory, the Qt translation
is used instead.