Make mkwrapper.pl write a banner to its generated files
In each generated file we want to have a banner that warns of
modification by hand and indicates the generator responsible
for the file. This patch extends mkwrapper.pl to write such
a header.

It also switches mkwrapper.pl away from barewords for file-handles.
Apparently those are no longer considered idiomatic Perl and they
broke the "iterate over list of handles" thing this patch does.