Fix overlay code related warnings in mumble
Apart from the usual shadow warnings and other spurious warnings
this code contained a lot of floating point conversions the compiler
was complaining about.

These are now either explicit or operate on double (whatever was
more convenient).

Even found a minor bug in OverlayClient::setupScene where the warning
indicated a typo in the rounding code.