Show Minimal View status in title
In the window title, indicate when Mumble is in Minimal View.

This should make it more clear,
especially if not connected to a server when there is *no* content in
the window, that it is not stuck/missing data.
This may also give the user a hint what to search for, if he is
unfamiliar with Minimal Mode.

This was suggested by the contributed usability suggestions from earlier this year.