Add support for moving multiple users/channels
Allow multi-select of users/channels (using ctrl/shift/mouse) and
subsequent drag-and-drop movement.

Previous behavior only allows movement of single user/channel
at a time, which is a hassle when dealing with many users.

This is strictly limited to drag-and-drop movement, and does not
allow multi-kick/ban/mute/message, etc.  That's a harder problem!