mumble.pro: add opus-src/celt and opus-src/include as -isystem headers.
This fixes the build on OS X.

Before this change, 3rdparty/celt-0.7.0-src/libcelt was marked
as a system include path.

Which should be fine.

However, with the aforementioned directory being treated as
a system directory, clang would handle an include statement
such as

    #include <celt.h>

by first checking the non-system directories.

In this case, it would pick opus-src/celt/celt.h.

This lead to bad things happening, such as missing
preprocessor defines, and missing typedefs for CELT's
integer types.

To work around that, mark Opus's header locations as
system include directories as well, so they're ordered
correctly when doing #include.